ci: simplify vitest setup
